requirements.txt
PyYAML
fire
polling
selenium
